CCollider Component
The CCollider component has been added. It is a simple wrapper around
the collision engine in the PaTFramework which allows you to add a
collider to a gameobject and render that collider. (This is currently
limited to the AABB and Sphere colliders in the PaTFramework). This
component will simply check for collisions whereas other components will
actually process them.