CMeshCollider added
The CMeshCollider has now been implemented into the collision engine. It
will allow you to construct a collider made from a mesh that has been
loaded into the CResourceBank object. This update also includes minor
changes to the CResourceList class as well as some changes to the Math
classes and functions.